requirements.txt
# EyeNet – Retinal Disease Detection & Analysis System
# Python dependencies

# ── Deep Learning ─────────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python-headless>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
pymongo>=4.6.0

# ── ML Utilities ─────────────────────────────────────────────────────────────
scikit-learn>=1.4.0
numpy>=1.26.0
pytest>=8.0.0

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.6.0

# ── Training extras ──────────────────────────────────────────────────────────
tqdm>=4.66.0
fpdf2>=2.7.0  